refactor: reduce global state
The default pattern of using spf13/cobra involves a lot of reliance upon
global variables for flags and options. This can cause side-effects,
especially when concurrency is involved, if one is not careful.

This change seeks to eliminate global state mutations and variables for
parsing flags and options from spf13/cobra into the other parts of my
helpers. We do this by providing an explicit demarcation point which
will help tests be written for more parts of these helpers in the
future.